+++ forge/changelog.md
@@ -1,1 +1,1 @@
-- Tweaked config values
+- Fixed reflected sounds